Refactor: move remaining site components
