+++ src/Sites/app/EmoteMenu/EmoteComponent.tsx
